requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
python-jose[cryptography] 
passlib[bcrypt]
python-multipart
httpx
email-validator